Create operating_system according to key and add casting to int in main_memory

+++ ciencia-da-computacao/bloco-37-estrutura-de-dados-i-arrays-listas-filas-e-pilhas/dia-01-arquitetura-de-computadores/pc/main_memory.py
@@ -7,7 +7,7 @@ class MainMemory:
 
     def get(self, index):
         try:
-            return self.loaded_memory[index]
+            return (int)(self.loaded_memory[index])
         except IndexError:
             return 0
